+++ changelog.md
@@ -11,6 +11,8 @@ Codion Change Log
 - TemporalField, temporalClass() renamed type().
 - NumberSpinnerBuilder, numberClass() renamed type().
 - NumberField, numberClass() renamed type().
+### is.codion.framework.domain.db
+- MetaDataColumn.columnClass() and columnName() renamed type() and name().
 
 ## 0.18.87
 ### is.codion.common.db